requirements.txt
aiogram>=3.4.0
web3>=6.15.0
py-clob-client==0.34.6
requests>=2.31.0
asyncpg>=0.29.0
aiohttp>=3.9.3
python-dotenv>=1.0.1
